pyproject.toml
[project]
name = "nonebot-plugin-pipe"
version = "0.1.0"
description = ""
authors = [{ name = "Jigsaw", email = "j1g5aw@foxmail.com" }]
dependencies = [
    "nonebot2<3.0.0,>=2.0.0",
    "nonebot-adapter-onebot>=2.2.0",
    "nonebot-plugin-localstore>=0.7.1",
    "nonebot-plugin-orm[default]>=0.7.5",
    "nonebot-plugin-user>=0.3.0",
]
requires-python = ">=3.9,<4.0"
readme = "README.md"
license = { text = "MIT" }

[tool.pdm.dev-dependencies]
pre-commit = ["ruff", "isort", "black", "nonemoji", "pre-commit"]
nonebot = ["fastapi", "uvicorn[standard]", "nb-cli"]

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = [
    "F",     # Pyflakes
    "W",     # pycodestyle warnings
    "E",     # pycodestyle errors
    "UP",    # pyupgrade
    "ASYNC", # flake8-async
    "C4",    # flake8-comprehensions
    "T10",   # flake8-debugger
    "T20",   # flake8-print
    "PYI",   # flake8-pyi
    "PT",    # flake8-pytest-style
    "Q",     # flake8-quotes
    "RUF",   # Ruff-specific rules
]
ignore = [
    "E402",   # module-import-not-at-top-of-file
    "UP037",  # quoted-annotation
    "RUF001", # ambiguous-unicode-character-string
    "RUF002", # ambiguous-unicode-character-docstring
    "RUF003", # ambiguous-unicode-character-comment
]

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.pyright]
pythonVersion = "3.9"
pythonPlatform = "All"
typeCheckingMode = "standard"
reportShadowedImports = false
disableBytesTypePromotions = true

[tool.nonebot]
adapters = [
    { name = "OneBot V12", module_name = "nonebot.adapters.onebot.v12" },
]
plugins = ["nonebot_plugin_pipe"]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"
